Allow Gateway IP to be parameterized for OSPF
The intent with this commit is to be capable of parameterizing the IP
that OSPF will use as a default gateway. This can be overridden in a
hackish way (by hardcoding a static IP later on), but this ultimately
needs to be fixed. This, however, will break backwards compatibility,
as it requires the `next_hop` parameter to be fed in.

Example interactive usage:

```
./cli.py -i nsx.ini esg -n example-edge routing_ospf \
--vnic_name uplink \
--vnic_ip 10.193.51.253 \
--area_id 200 \
--auth_type password \
--auth_value yolo \
--next_hop 10.193.51.1
```

Expected STDOUT:

```
Edge Service Gateway example-edge configured for OSPF
```